% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{FindPatient}
\alias{FindPatient}
\title{Matches by EVERY thing. If you want by PHN submit a patient with only PHN}
\usage{
FindPatient(patient, tolerance = 0)
}
\description{
Matches by EVERY thing. If you want by PHN submit a patient with only PHN
}
