% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physicians.R
\name{FindPhysician}
\alias{FindPhysician}
\title{Matches by EVERY thing. If you want by userName submit a physicain with only userName}
\usage{
FindPhysician(physician, tolerance = 0)
}
\description{
Matches by EVERY thing. If you want by userName submit a physicain with only userName
}
