% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{FindPatient}
\alias{FindPatient}
\title{Matches by EVERY thing. If you want by Code submit a patient with only Code}
\usage{
FindPatient(patient, tolerance = 1)
}
\description{
Matches by EVERY thing. If you want by Code submit a patient with only Code
}
